refactor(ui): Phase 5 — Zustand persist middleware + constants split
Stores:
- use-auth-store: manual localStorage → persist() with partialize
  (token, userId, senderID)
- use-ui-store: manual localStorage → persist() with partialize
  (theme, language, timezone, sidebarCollapsed)
- use-team-event-store: manual persist helpers → persist() with
  partialize (last 20 events), onRehydrateStorage seeds counter

Constants:
- Split lib/constants.ts (144 lines) into:
  - lib/routes.ts (46 lines) — route paths
  - lib/timezone-utils.ts (74 lines) — IANA timezone data
  - lib/constants.ts (30 lines) — barrel re-export

ForwardRef: skipped — all UI primitives already use React 19
ref-as-prop pattern.

No behavior changes. Build passes.
