+++ ui/web/src/lib/routes.ts
@@ -1,0 +1,46 @@
+export const ROUTES = {
+  LOGIN: "/login",
+  OVERVIEW: "/overview",
+  CHAT: "/chat",
+  CHAT_PATTERN: "/chat/:sessionKey?",
+  AGENTS: "/agents",
+  AGENT_DETAIL: "/agents/:id",
+  AGENT_CODEX_POOL: "/agents/:id/codex-pool",
+  IMPORT_EXPORT: "/import-export",
+  SESSIONS: "/sessions",
+  SESSION_DETAIL: "/sessions/:key",
+  SKILLS: "/skills",
+  SKILL_DETAIL: "/skills/:id",
+  CRON: "/cron",
+  CRON_DETAIL: "/cron/:id",
+  CONFIG: "/config",
+  TRACES: "/traces",
+  TRACE_DETAIL: "/traces/:id",
+  EVENTS: "/events",
+  USAGE: "/usage",
+  CHANNELS: "/channels",
+  CHANNEL_DETAIL: "/channels/:id",
+  CONTACTS: "/contacts",
+  APPROVALS: "/approvals",
+  NODES: "/nodes",
+  LOGS: "/logs",
+  PROVIDERS: "/providers",
+  PROVIDER_DETAIL: "/providers/:id",
+  TEAMS: "/teams",
+  TEAM_DETAIL: "/teams/:id",
+  BUILTIN_TOOLS: "/builtin-tools",
+  CLI_CREDENTIALS: "/cli-credentials",
+  MCP: "/mcp",
+  TTS: "/tts",
+  STORAGE: "/storage",
+  PENDING_MESSAGES: "/pending-messages",
+  MEMORY: "/memory",
+  KNOWLEDGE_GRAPH: "/knowledge-graph",
+  ACTIVITY: "/activity",
+  API_KEYS: "/api-keys",
+  PACKAGES: "/packages",
+  TENANTS: "/admin/tenants",
+  TENANT_DETAIL: "/admin/tenants/:id",
+  SETUP: "/setup",
+  SELECT_TENANT: "/select-tenant",
+} as const;
